Add WriteUp class that will contain each write-up
Implemented the writeup class with a constructor that parses the
standard organization of the CERNLIN short write-ups. The class now
knows what title, author, ID, keywords, routines, library, programming
language, submitter, and sumbission and revision dates to associate with
the writeup.

I am following test driven design. The test directory contains the test
implementations and the data needed for the tests.